+++ servicetalk-http-api/src/main/java/io/servicetalk/http/api/DefaultStreamingContentCodingBuilder.java
@@ -1,0 +1,43 @@
+/*
+ * Copyright © 2020 Apple Inc. and the ServiceTalk project authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.servicetalk.http.api;
+
+abstract class DefaultStreamingContentCodingBuilder implements StreamingContentCodingBuilder {
+
+    private static final int CHUNK_SIZE = 1 << 10; //1KiB
+
+    protected int maxAllowedPayloadSize = DEFAULT_MAX_ALLOWED_COMPRESSED_PAYLOAD;
+
+    @Override
+    public StreamingContentCodingBuilder setMaxAllowedPayloadSize(final int maxAllowedPayloadSize) {
+        this.maxAllowedPayloadSize = maxAllowedPayloadSize;
+        return this;
+    }
+
+    static class GzipStreamingContentCodingBuilder extends DefaultStreamingContentCodingBuilder {
+        @Override
+        public StreamingContentCoding build() {
+            return new GzipContentCoding(CHUNK_SIZE, maxAllowedPayloadSize);
+        }
+    }
+
+    static class DeflateStreamingContentCodingBuilder extends DefaultStreamingContentCodingBuilder {
+        @Override
+        public StreamingContentCoding build() {
+            return new DeflateContentCoding(CHUNK_SIZE, maxAllowedPayloadSize);
+        }
+    }
+}
